Implementación de CardiovascularMetrics y MetabolicMetrics

+++ src/main/java/healthcalc/GUI/calculadora_vista.java
@@ -274,7 +274,7 @@ public class calculadora_vista {
                 JOptionPane.ERROR_MESSAGE);
     }
 
-    public void setResultText(float resultado2) {
+    public void setResultText(double resultado2) {
         resultado.setText(String.valueOf(resultado2));
     }
